Split library into jsearch-core and jsearch-examples.
The NPuzzle problem is intended as an example on how to use JSearch. It should be possible to import JSearch without these example files, which is why it was separated to a different module. The Dummy problem (which is typically only useful for the tests), has been moved to the test folder.